% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.data.frame}
\alias{st_as_sf.sf}
\title{Convert foreign object to an sf object}
\usage{
st_as_sf(x, ...)

\method{st_as_sf}{data.frame}(x, ..., agr = NA_agr_, coords, wkt,
  dim = "XYZ", remove = TRUE)

\method{st_as_sf}{sf}(x, ...)
}
\arguments{
\item{x}{object to be converted into an object class \code{sf}}

\item{...}{passed on to \link{st_sf}, might included crs}

\item{agr}{character vector; see details section of \link{st_sf}}

\item{coords}{in case of point data: names or numbers of the numeric columns holding coordinates}

\item{wkt}{name or number of the character column that holds WKT encoded geometries}

\item{dim}{passed on to \link{st_point} (only when argument coords is given)}

\item{remove}{logical; when coords or wkt is given, remove these columns from data.frame?}
}
\description{
Convert foreign object to an sf object
}
\details{
setting argument \code{wkt} annihilates the use of argument \code{coords}. If \code{x} contains a column called "geometry", \code{coords} will result in overwriting of this column by the \link{sfc} geometry list-column.  Setting \code{wkt} will replace this column with the geometry list-column, unless \code{remove_coordinates} is \code{FALSE}.
}
\examples{
pt1 = st_point(c(0,1))
pt2 = st_point(c(1,1))
st_sfc(pt1, pt2)
d = data.frame(a = 1:2)
d$geom = st_sfc(pt1, pt2)
df = st_as_sf(d)
d$geom = c("POINT(0 0)", "POINT(0 1)")
df = st_as_sf(d, wkt = "geom")
d$geom2 = st_sfc(pt1, pt2)
st_as_sf(d) # should warn
data(meuse, package = "sp")
meuse_sf = st_as_sf(meuse, coords = c("x", "y"), crs = 28992, agr = "constant")
meuse_sf[1:3,]
summary(meuse_sf)
}

